app/requirements.txt
fastapi
jinja2
python-jose
sqlalchemy
uvicorn
pymysql
python-multipart
bcrypt
python-dotenv